remove SparkCatalogShim
